使用 Wrapper 进行规范 Entity

+++ src/IdS4/IdS4.Infrastructure/Repositories/LogsRepository.cs
@@ -6,7 +6,7 @@ using RajsLibs.Repository.EfCore;
 
 namespace IdS4.Infrastructure.Repositories
 {
-    public class LogsRepository : RepositoryBase<IdS4LogDbContext, Log, int>, ILogRepository
+    public class LogsRepository : RepositoryBase<IdS4LogDbContext, Log, long>, ILogRepository
     {
         public LogsRepository(IEfUnitOfWork<IdS4LogDbContext> unitOfWork) 
             : base(unitOfWork)
