+++ src/IdS4/IdS4/Repositories/ILogRepository.cs
@@ -3,7 +3,7 @@ using RajsLibs.Repositories;
 
 namespace IdS4.Repositories
 {
-    public interface ILogRepository : IRepository<Log, int>
+    public interface ILogRepository : IRepository<Log, long>
     {
 
     }
